Refactor: Split boat hire logic out of Vehicle.cpp
Part of #25951

Moved all boat hire functions from Vehicle.cpp into a dedicatedVehicle.BoatHire.cpp file. Functions are moved intact with nologic changes.

Moved functions:
- UpdateTravellingBoatHireSetup
- UpdateDepartingBoatHire
- UpdateTravellingBoat
- TryReconnectBoatToTrack
- UpdateMotionBoatHire
- UpdateBoatLocation
- vehicle_boat_is_location_accessible
