+++ src/openrct2/libopenrct2.vcxproj
@@ -1121,6 +1121,7 @@
     <ClCompile Include="ride\TrackPaint.cpp" />
     <ClCompile Include="ride\TrackStyle.cpp" />
     <ClCompile Include="ride\TrainManager.cpp" />
+    <ClCompile Include="ride\Vehicle.BoatHire.cpp" />
     <ClCompile Include="ride\Vehicle.cpp" />
     <ClCompile Include="ride\VehicleData.cpp" />
     <ClCompile Include="ride\VehicleGeometry.cpp" />
